🐛 Define `luasnip` by `require`-ing it
There was a problem: when Tab was pressed while there was no suggestion,
it would not indent the line, instead it would show errors. This was due
to `luasnip` being undefined.

Now `luasnip` is defined as `require('luasnip')`, which fixes the issue.

So, the changes in f4b232c are reverted since they are no longer needed.